requirements.txt
click
CMake
dlib
enum-compat
enum34
eventlet
Flask
flask-cors
Flask-SocketIO
flask_sqlalchemy
gevent
gevent-websocket
greenlet
gunicorn
imutils
itsdangerous
Jinja2
MarkupSafe
numpy
olefile
opencv-python
opencv-contrib-python
Pillow
pdf2image
pybase64
python-engineio
python-socketio
scikit-learn
six
Werkzeug

